+++ crates/hyperqueue/src/server/state.rs
@@ -252,7 +252,12 @@ impl State {
             .on_worker_added(msg.worker_id, msg.configuration);
     }
 
-    pub fn process_worker_lost(&mut self, msg: LostWorkerMessage) {
+    pub fn process_worker_lost(
+        &mut self,
+        state_ref: &StateRef,
+        tako_ref: &Backend,
+        msg: LostWorkerMessage,
+    ) {
         log::debug!("Worker lost id={}", msg.worker_id);
         let worker = self.workers.get_mut(&msg.worker_id).unwrap();
         worker.set_offline_state(msg.reason.clone());
@@ -261,9 +266,19 @@ impl State {
             autoalloc.on_worker_lost(msg.worker_id, &worker.configuration, msg.reason.clone());
         }
 
+        let mut tasks_to_cancel: Map<JobId, Vec<TakoTaskId>> = Map::new();
         for task_id in msg.running_tasks {
             let job = self.get_job_mut_by_tako_task_id(task_id).unwrap();
-            job.set_waiting_state(task_id);
+            job.increment_task_crash_count(task_id);
+            if job.is_often_crashing_task(task_id) {
+                tasks_to_cancel.entry(job.job_id).or_default().push(task_id);
+            } else {
+                job.set_waiting_state(task_id);
+            }
+        }
+
+        for (job_id, task_ids) in tasks_to_cancel {
+            cancel_tasks_from_callback(state_ref, tako_ref, job_id, task_ids);
         }
 
         self.event_storage.on_worker_lost(msg.worker_id, msg.reason);